Visualization and Export
This commit included the new tabbed layout with active tabs for
"visualization" and "export" options. The export option allows the user
to save the active layer to .kml. There are buttons for track details
and google earth, but they are not yet functional. The "visualization"
tab allows the user to apply camera views to one or many points in the
track.

The edit session must be used to change the camera options. Also, the
user must "cancel" changes to the selected layer to rollback the
changes, rather than just using "rollback". This may be a bug, I need to
check into this.

Also, improvements have been made to the import tab. A check box for
"head of track" was added to draw a marker for the start point of the
track. A "sync" button was added to find the point in the track where
the audio starts and draw a red marker. Both markers are drawn "in
memory".