Rendering; IconStyle & LabelStyle
Styling of individual points using the kml IconStyle and LabelStyle
tags. User should select points from the active layer in the table of
contents and then apply styles to the point icons/labels. hotspot tag is
not enabled yet. transparency tag is only enabled if the user selects a
color. The icon and label styles can be applied independently.

Also, support for kmz export was added.

To Do:
1) Add point scaling based on attributes.
2) Add support for <Model> tag using DAE files