KML Point Models
This commit focused on adding functionality to the rendering tab.
Specifically, adding kml <Model> functionality. Applying models to
selected points is the same workflow as the "IconStyle" section. Select
a point, or all the points, and browse to a Collada .dae file. The
default values will put the model at the location of the points at
ground level.