Audio Export to .kml/.kmz
Support was added for exporting an audio (.wav) file to the .kml/.kmz
output of the GPS track. The .wav file must be in the YYYYMMDDHHMMSS.wav
format. If there is a time delay from the GPS track and the audio, Milk
Machine will automatically populate the <gx:delayedStart> kml tag.