'camera' field coding
Bug Fix:   The 'camera' field exceeds the maximum character limit of 254
sometimes. This means that QGIS cuts of extra characters when saving as
a shapefile. However, during and edit session the field will hold the
extra characters. This allows a safe export in MM. There is no error in
QGIS for the extra characters. The solution is to code all of the
parameters. This brings the camera field down to about 180 characters.
Bug Fix: The 'tilt' field needs only to autopopulate when it is empty,
not when the user enters a value. This stops it from overwriting the
user values.