<LookAt> and Mac Formatting
Task:  Create a <LookAt> tab in the Tour tab of Milk Machine. The
<LookAt> tab has two modes; "custom" and "circle around". Both
automatically calculate latitude and longitude. In custom mode, the user
can select 1 or more points and apply a custom <LookAt> to each. If
multiple points are selected, lat/long for each <LookAt> will be the
lat/long of that point/model. In "circle around" mode, 1 or more points
can also be selected. If multiple points are selected, MM calculates the
geometric centroid of the selected points for the <LookAt> lat/long. If
only 1 point is selected, it's lat/long will be the lat/long for the
<LookAt>. "circle around" also has some parameter for customizing the
number of rotations, starting heading, and duration.