Formatting and numpy filtering fixes
Bug Fix: Size policy changes for labels and buttons. Hoping that "fixed"
"fixed" works
Bug Fix: The matplotlib library is not that same as the one on Windows,
so no plotting!!. Only works on windows.
Bug Fix: numpy on Mac is an older version, so I can't call scipy/numpy
as pylab. Call as numpy.