Circle Improvement
Bug Fix: Change the duration calculation in "circle around" to be the
time difference between the start and the end of the selected points
Task:  On "editcomplete" signal for any line edit in "circle around",
the "duration" will recalculate (just in case the selection was
changed). Also on "apply"
Task:  Provide user with a warning when overwriting an existing circle
around with another circle around. The user will be prompted to either
overwrite the existing lookat cells, or to cancel. The conflicting
circle around will be deleted.