Spline Interpolation
Task:  Removed the "quadratic" interpolation from the filtering options.
Replaced it with a weighted spline interpolator that accepts
"Quadratic", "Cubic", 4th, and 5th order polynomial smoothers. The
operates by regressing the Y axis (latitude) on the X axis (longitude),
which means that X is the predictor and y is the dependent variable.