Prio A and B v2
Task: On "Sync" button. The user will be informed if the audio lines up
with the track start time, or not.
Task: On "Sync" button. Added the "Indicate Sync Point" checkbox so that
the user can control if the "sync" point is shown or not. This should
reduce repetitive layers being drawn.
Task: Package mutagen and gpxpy now ships with MilkMachine
Task: Full support for .mp3 file integration.
Task: Follow behind camera offset. This was first called "swirl" by
Esther. The way this works is to apply a constant offset (in degrees) to
the heading value. The heading value is calculated by MM, the offset is
then added to each calculated heading (0-360)