Prio A and B v3 bug fixes v1
Bug Fix: Export with .mp3 fails. This was due to MM still trying to use
the .wav library. Changed to mutagen, and export works.
Bug Fix: When the audio and video have matching start times, the dialog
box says "False". The variable "matchdict" was not being updated when
the first row was a match.